requirements.txt
python-dotenv==1.0.1
langchain==0.2.7
langchain-community==0.2.7
langchain-openai==0.1.16